Split siteWrapper into multiple files
