+++ appservice/src/index.ts
@@ -3,10 +3,16 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-export * from './SiteWrapper';
 export * from './createAppService/createFunctionApp';
 export * from './createAppService/createWebApp';
+export * from './deploy/deploy';
+export * from './deleteSite';
+export * from './editScmType';
+export * from './getFile';
+export * from './getKuduClient';
+export * from './pingFunctionApp';
+export * from './putFile';
+export * from './SiteClient';
+export * from './startStreamingLogs';
 export * from './tree/AppSettingsTreeItem';
 export * from './tree/AppSettingTreeItem';
-export * from './ILogStream';
-export * from './IFileResult';
