+++ appservice/src/signRequest.ts
@@ -1,0 +1,18 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { ServiceClientCredentials, WebResource } from "ms-rest";
+
+export async function signRequest(req: WebResource, cred: ServiceClientCredentials): Promise<void> {
+    await new Promise((resolve: () => void, reject: (err: Error) => void): void => {
+        cred.signRequest(req, (err: Error) => {
+            if (err) {
+                reject(err);
+            } else {
+                resolve();
+            }
+        });
+    });
+}
